[fix] 테스트 코드 폴더 위치 수정
