+++ src/main/java/sopt/jeolloga/domain/templestay/api/controller/FilterController.java
@@ -1,43 +1,21 @@
 package sopt.jeolloga.domain.templestay.api.controller;
 
-import jakarta.servlet.Filter;
-import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 import sopt.jeolloga.domain.member.core.exception.MemberCoreException;
 import sopt.jeolloga.domain.templestay.api.dto.*;
 import sopt.jeolloga.domain.templestay.api.service.FilterService;
-import sopt.jeolloga.domain.templestay.api.service.FilterServiceV1;
-import sopt.jeolloga.domain.templestay.api.service.FilterServiceV2;
-import sopt.jeolloga.domain.templestay.api.service.TemplestaySearchService;
 import sopt.jeolloga.exception.ErrorCode;
-
-import java.util.List;
 
 @RestController
 public class FilterController {
 
     private final FilterService filterService;
 
-    public FilterController(FilterService filterService,  FilterServiceV1 filterServiceV1, FilterServiceV2 filterServiceV2) {
+    public FilterController(FilterService filterService) {
         this.filterService = filterService;
     }
-
-//    @GetMapping("/public/filter")
-//    public ResponseEntity<FilterRes> getFilters() {
-//
-//        FilterRes filterList = filterService.getFilters();
-//        return ResponseEntity.ok(filterList);
-//    }
-
-//    @GetMapping("/public/filter/reset")
-//    public ResponseEntity<ResetFilterRes> getResetFilter() {
-//
-//        ResetFilterRes resetFilterRes = filterService.getFilterReset();
-//        return ResponseEntity.ok(resetFilterRes);
-//    }
 
     @PostMapping("public/filter/count")
     public ResponseEntity<FilterCountRes> getFilteredTemplestayNum(@RequestBody FilterReq filter){
