+++ src/test/java/performance/testCode/FilterCountV1Test.js
@@ -2,26 +2,25 @@ import http from 'k6/http';
 import { check, sleep } from 'k6';
 
 export let options = {
-  vus: 100,         // 동시 사용자 100명
-  duration: '10s'   // 10초 동안 테스트
+  vus: 10,         // 동시 사용자 10명
+  duration: '600s'  // 10분 동안 테스트
 };
 
 export default function () {
-
-  let url = "http://127.0.0.1:8080/public/search/v2?page=1&userId=2";
+  let url = 'http://127.0.0.1:8080/public/filter/count/v1'; // API 엔드포인트
 
   let payload = JSON.stringify({
-     "content" : "",
+     "content" : " 사",
      "region": {
        "강원": 0,
-       "경기": 0,
-       "경남": 0,
-       "경북": 0,
+       "경기": 1,
+       "경남": 1,
+       "경북": 1,
        "광주": 0,
        "대구": 0,
        "대전": 0,
        "부산": 0,
-       "서울": 0,
+       "서울": 1,
        "인천": 0,
        "전남": 0,
        "전북": 0,
@@ -30,25 +29,25 @@ export default function () {
        "충북": 0
      },
      "type": {
-       "당일형": 0,
-       "휴식형": 0,
-       "체험형": 0
+       "당일형": 1,
+       "휴식형": 1,
+       "체험형": 1
      },
      "purpose": {
-       "힐링": 0,
+       "힐링": 1,
        "전통문화 체험": 0,
-       "심신치유": 0,
+       "심신치유": 1,
        "자기계발": 0,
        "여행 일정": 0,
-       "사찰순례": 0,
-       "휴식": 0,
+       "사찰순례": 1,
+       "휴식": 1,
        "호기심": 0
      },
      "activity": {
-       "발우공양": 0,
-       "108배": 0,
-       "스님과의 차담": 0,
-       "등산": 0,
+       "발우공양": 1,
+       "108배": 1,
+       "스님과의 차담": 1,
+       "등산": 1,
        "새벽 예불": 0,
        "사찰 탐방": 0,
        "염주 만들기": 0,
@@ -57,23 +56,23 @@ export default function () {
        "명상": 0,
        "산책": 0,
        "요가": 0,
-       "기타": 0
+       "기타": 1
      },
      "price": {
        "minPrice": 0,
-       "maxPrice": 300000
+       "maxPrice": 200000
      },
      "etc": {
        "절밥이 맛있는": 0,
        "TV에 나온": 0,
-       "연예인이 다녀간": 0,
-       "근처 관광지가 많은": 0,
-       "속세와 멀어지고 싶은": 0,
+       "연예인이 다녀간": 1,
+       "근처 관광지가 많은": 1,
+       "속세와 멀어지고 싶은": 1,
        "단체 가능": 0,
-       "동물 친구들과 함께": 0,
-       "유튜브 운영 중인": 0
+       "동물 친구들과 함께": 1,
+       "유튜브 운영 중인": 1
      }
-  });
+    });
 
   let params = {
     headers: { "Content-Type": "application/json" }
