+++ src/test/java/sopt/jeolloga/JeollogaApplicationTests.java
@@ -3,6 +3,7 @@ package sopt.jeolloga;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
+//@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @SpringBootTest
 class JeollogaApplicationTests {
